decompressor: eliminate empty frame optimization so `allow_extra_data=False` is respected
In the beginning we had an optimization to handle the special case
of a 0 length zstd frame. We later added the `allow_extra_data`
argument. Unfortunately, we failed to account for the interaction
between this optimization and `allow_extra_data`.

This commit fixes the logical error and enables `allow_extra_data=False`
to work in the presence of an empty frame.